updated documentation and added unit tables

+++ build_doc.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
+python3 build_tables.py
 cd docs
-make clean
-rm -fr ./source/api
+if [[ "$1" != "-u" ]]; then
+    make clean
+    rm -fr ./source/api
+fi
 sphinx-apidoc -o ./source/api ../src/*
 sphinx-build -b html source build/html -v
 cd build/html
